+++ pkg/events/pubsub.go
@@ -15,10 +15,9 @@
 package events
 
 import (
-	"runtime/trace"
-	"sync"
+	"context"
 
-	"github.com/gobwas/glob"
+	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 )
 
 // PubSub interface combines the Publisher and Subscriber interfaces.
@@ -35,97 +34,19 @@ type Publisher interface {
 
 // Subscriber interface lets you subscribe to events.
 type Subscriber interface {
-	// Subscribe adds an event handler to the default event pubsub.
+	// Subscribe to events that match the name and identifiers.
+	// The subscription continues until the context is canceled.
 	// The name can be a glob in order to catch multiple event types.
+	// If supplying identifiers, then only events matching _any_ of those
+	// identifiers will get sent to the handler.
 	// The handler must be non-blocking.
-	Subscribe(name string, hdl Handler) error
-	// Unsubscribe removes an event handler from the default event pubsub.
-	// Queued or in-transit events may still be delivered to the handler
-	// even after Unsubscribe returns.
-	Unsubscribe(name string, hdl Handler)
+	Subscribe(ctx context.Context, name string, identifiers []*ttnpb.EntityIdentifiers, hdl Handler) error
 }
 
-type handler struct {
-	eventName string
-	glob.Glob
-	Handler
-}
-
-type tracedEvent struct {
-	event Event
-	trace *trace.Region
-}
-
-type pubsub struct {
-	mu       sync.RWMutex
-	handlers []handler
-	events   chan tracedEvent
-}
-
-// DefaultBufferSize is the default number of events that can be buffered before Publish starts to block.
-const DefaultBufferSize = 64
-
-// NewPubSub returns a new event pubsub and starts a goroutine for handling.
-func NewPubSub(bufSize uint) PubSub {
-	e := &pubsub{
-		events: make(chan tracedEvent, bufSize),
-	}
-	go e.Run()
-	return e
-}
-
-func (e *pubsub) Run() {
-	for evt := range e.events {
-		e.mu.RLock()
-		handlers := e.handlers
-		e.mu.RUnlock()
-		for _, l := range handlers {
-			if l.Match(evt.event.Name()) {
-				l.Notify(evt.event)
-			}
-		}
-		evt.trace.End()
-	}
-}
-
-func (e *pubsub) Subscribe(name string, hdl Handler) error {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	for _, l := range e.handlers {
-		if l.eventName == name && l.Handler == hdl {
-			return nil
-		}
-	}
-	glob, err := glob.Compile(name, '.')
-	if err != nil {
-		return err
-	}
-	e.handlers = append(e.handlers, handler{
-		eventName: name,
-		Glob:      glob,
-		Handler:   hdl,
-	})
-	subscriptions.WithLabelValues(name).Inc()
-	return nil
-}
-
-func (e *pubsub) Unsubscribe(name string, hdl Handler) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	for i, l := range e.handlers {
-		if l.eventName == name && l.Handler == hdl {
-			e.handlers = append(e.handlers[:i], e.handlers[i+1:]...)
-			subscriptions.WithLabelValues(name).Dec()
-			return
-		}
-	}
-}
-
-func (e *pubsub) Publish(evt Event) {
-	localEvent := local(evt)
-	publishes.WithLabelValues(evt.Name()).Inc()
-	e.events <- tracedEvent{
-		event: localEvent.withCaller(),
-		trace: trace.StartRegion(evt.Context(), "publish event"),
-	}
+// Subscription is the interface for PubSub subscriptions.
+type Subscription interface {
+	// Match returns whether the event matches the subscription.
+	Match(Event) bool
+	// Notify notifies the subscription of a new matching event.
+	Notify(Event)
 }