+++ CHANGELOG.md
@@ -15,6 +15,9 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Changed
 
+- Changed the pub/sub channels that the Redis backend of the Events system uses.
+- Changed the encoding of events transported by the Redis backend of the Events system.
+
 ### Deprecated
 
 ### Removed